fraud-detection/requirements.txt
streamlit
python-dotenv==0.21.0
azure-storage-blob==12.8.1